requirements.txt
Flask
Flask-MySQLdb
Flask-SQLAlchemy
SQLAlchemy
mysql-connector-python
requests
python-dotenv
importlib-resources
Flask-WTF
WTForms
email-validator
pyjwt
pyjwt[EdDSA]
bcrypt
flask-cors
flask-sslify
werkzeug
unidecode
boto3
elasticsearch
Gensim==4.1.0
python-telegram-bot